Separate TestRandomGenerator from utility_lib
