+++ test/test_common/BUILD
@@ -113,6 +113,15 @@ envoy_cc_test_library(
 )
 
 envoy_cc_test_library(
+    name = "test_random_generator_lib",
+    srcs = ["test_random_generator.cc"],
+    hdrs = ["test_random_generator.h"],
+    deps = [
+        "//source/common/common:utility_lib",
+    ],
+)
+
+envoy_cc_test_library(
     name = "utility_lib",
     srcs = ["utility.cc"],
     hdrs = ["utility.h"],
@@ -124,6 +133,7 @@ envoy_cc_test_library(
         ":logging_lib",
         ":printers_lib",
         ":resources_lib",
+        ":test_random_generator_lib",
         ":test_time_lib",
         ":thread_factory_for_test_lib",
         "//envoy/buffer:buffer_interface",
@@ -187,6 +197,7 @@ envoy_cc_test_library(
     srcs = ["file_system_for_test.cc"],
     hdrs = ["file_system_for_test.h"],
     deps = [
+        ":simulated_time_system_lib",
         "//source/common/common:utility_lib",
         "//source/common/filesystem:filesystem_lib",
     ],
@@ -273,8 +284,9 @@ envoy_cc_test_library(
     srcs = ["simulated_time_system.cc"],
     hdrs = ["simulated_time_system.h"],
     deps = [
+        ":test_random_generator_lib",
         ":test_time_system_interface",
-        ":utility_lib",
+        ":thread_factory_for_test_lib",
         "//source/common/event:event_impl_base_lib",
         "//source/common/event:real_time_system_lib",
         "//source/common/event:timer_lib",
