+++ test/test_common/simulated_time_system_test.cc
@@ -6,7 +6,6 @@
 #include "test/mocks/common.h"
 #include "test/mocks/event/mocks.h"
 #include "test/test_common/simulated_time_system.h"
-#include "test/test_common/utility.h"
 
 #include "event2/event.h"
 #include "gtest/gtest.h"
