+++ tools/code_format/config.yaml
@@ -147,10 +147,10 @@ paths:
     - test/integration/integration.h
     - test/test_common/simulated_time_system.cc
     - test/test_common/simulated_time_system.h
+    - test/test_common/test_random_generator.cc
     - test/test_common/test_time.cc
     - test/test_common/test_time.h
     - test/test_common/utility.cc
-    - test/test_common/utility.h
     - test/tools/wee8_compile/wee8_compile.cc
 
   # Tests in these paths may make use of the Registry::RegisterFactory constructor or the
